FaST-LMM full rank

We first consider the scenario where $K$ is defined by a square semi-positive definite matrix and $K_{22}$ has been computed by project_orth_covar. Following Lippert et al. (2011), we call this the "full rank" scenario, although neither $K$ nor $K_{22}$ actually has to have full rank.

The spectral decomposition of $K_{22}$ can be written as $K_{22}=U \Lambda U^T$, with $U\in\mathbb{R}^{(n-d)\times (n-d)}$ unitary and $\Lambda\in\mathbb{R}^{(n-d)\times (n-d)}$ diagonal with non-negative diagonal elements $\lambda_i=\Lambda_{ii}\geq 0$, which we assume are ordered, $\lambda_1\geq\lambda_2\geq \dots \geq \lambda_{n-d}\geq 0$

The columns of $U$ are the eigenvectors of $K_{22}$, and we denote them as $u_i \in\mathbb{R}^{n-d}$, with $K_{22} u_i=\lambda_i u_i$. The matrix $U$ can be used to rotate the data $y_2$:

\[\tilde{y} = U^T y_2 \in \mathbb{R}^{n-d}\]

with components

\[\tilde{y}_i = \langle u_i,y_2\rangle,\quad i=1,\dots,n-d\]

These results allow to express the terms in $\mathcal{\ell}_R$

\[\begin{aligned} \frac1{n-d}\log\det(K_{22}+\delta I) &= \frac1{n-d}\sum_{i=1}^{n-d} \log(\lambda_i+\delta)\\ \hat\sigma^2 = \frac1{n-d}\langle y_2,(K_{22}+\delta I)^{-1} y_2\rangle &= \frac1{n-d}\sum_{i=1}^{n-d} \frac1{\lambda_i+\delta} \langle y_2,u_i\rangle^2 = \frac1{n-d}\sum_{i=1}^{n-d} \frac{\tilde{y}_i}{\lambda_i+\delta} \end{aligned}\]

The values of $\hat\sigma^2$ and $\ell_R$ as a function of the parameter $\delta$ and vectors $\lambda$ and $\tilde y$ are computed by the functions

FaSTLMMlight.sigma2_mle_fullrank — Function

sigma2_mle_fullrank(δ, λ, yr)

Compute the MLE of the variance parameter given the variance ratio δ, the eigenvalues of the kernel matrix, and the rotated response vector.

FaSTLMMlight.minus_log_like_fullrank — Function

minus_log_like_fullrank(δ, λ, yr)

Compute the minus log-likelihood of the model, scaled by the number of samples and without constant factors, given the variance ratio δ, the eigenvalues of the kernel matrix, the rotated response vector and (optionally) the rotated covariates.

The function

FaSTLMMlight.fastlmm_fullrank — Function

fastlmm_fullrank(y,K; covariates = [], mean = true)

For a linear mixed model / Gaussian process,

\[y \sim N\bigl(X\beta, \sigma^2(K + \delta I) \bigr)\]

where $y$ is the response vector, $X$ is a matrix of covariates, and $K$ is a full-rank kernel matrix, compute the restricted maximum-likelihood estimates (REMLs) of the variance parameter $\sigma^2$ and the variance ratio $\delta$ using FaST-LMM with a full-rank kernel matrix. Compared to the original FaST-LMM algorithm, we first project out the (optional) covariates, incl. an (optional) constant off-set (mean=true), from the response vector and the kernel matrix. This avoids all matrix computations in the variance parameter estimation. Estimates for the fixed effects $\beta$ are not computed.

FaSTLMMlight.delta_mle_fullrank — Function

delta_mle_fullrank(λ, yr)

Compute the MLE of the variance ratio δ given the eigenvalues of the kernel matrix and the rotated response vector by solving the non-convex optimization problem formulated in the FaST-LMM paper.

FaSTLMMlight.beta_mle_fullrank_lazy — Function

beta_mle_fullrank_lazy(y, K, X, σ², δ)

Lazy implementation of the MLE of the fixed effects weights given the response vector y, the kernel matrix K, the covariates X, the variance parameter σ² and the variance ratio δ. This function does not use spectral factorization, and should only be applied once, using the MLEs of the variance parameters.

FaSTLMMlight.softplus — Function

softplus(x)

Compute the softplus function, i.e. log(1 + exp(x)), element-wise.